provide buttons to toggle the closed and locked states. and buttons' text changes to reflect the state the door will be in if clicked
