+++ src/controls/Controls.spec.js
@@ -1,1 +1,20 @@
-// Test away!
+// Test away
+
+import React from 'react';
+import { render, fireEvent } from '@testing-library/react';
+
+import Controls from './Controls';
+
+/*
+test('provide buttons to toggle the closed and locked states.', () => {
+  const { getByText } = render(<Controls />);
+
+  getByText(/close/i);
+  const closeButton = getByText(/close gate/i);
+  fireEvent.click(closeButton);
+  getByText(/open/i);
+
+
+  //getByText(/open/i);
+});
+MAKES MORE SENCE TO DO FROM DISPAY*/
